trie: missing type conversion before abs()

+++ src/libknot/internal/trie/hat-trie.c
@@ -500,7 +500,7 @@ int hattrie_split_mid(node_ptr node, unsigned *left_m, unsigned *right_m)
 
     while (j + 1 < node.b->c1) {
         d = abs((int) (*left_m + cs[j + 1]) - (int) (*right_m - cs[j + 1]));
-        if (d <= abs(*left_m - *right_m) && *left_m + cs[j + 1] < all_m) {
+        if (d <= abs((int) (*left_m) - (int) (*right_m)) && *left_m + cs[j + 1] < all_m) {
             j += 1;
             *left_m  += cs[j];
             *right_m -= cs[j];
